% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Create a plotly heatmap of habitat changes}
\usage{
make_heatmap(df, species = "", title = "")
}
\arguments{
\item{species}{The name of the species to plot}

\item{data}{The data set to subset}
}
\value{
A plotly heatmap
}
\description{
Create a plotly heatmap of habitat changes
}
\examples{
\dontrun{
make_heatmap(all_sp, "RiceRat")
}
}

